refactor(npc): break npc.py monolith into src/npc/ package with mixins
Mirrors the src/player/ pattern. The 1 757-line npc.py is replaced by a
structured directory package:

  src/npc/
    __init__.py      — public re-exports (zero import changes needed)
    _combat.py       — NPCCombatMixin (select_move, add_move, reset_combat_moves, combat_engage)
    _loot.py         — NPCLootMixin (die, before_death, drop_inventory, roll_loot)
    _shop.py         — MerchantShopMixin (moved from src/npc_shop_mixin.py)
    _llm.py          — MynxLLMMixin (moved from src/npc_mynx_mixin.py; __file__ depth fixed)
    _base.py         — NPC (inherits both mixins + Combatant) and Friend
    _enemies.py      — Slime, Testexp, RockRumbler, Lurker, GiantSpider,
                       CaveBat, ElderSlime, KingSlime, StatusDummy
    _merchants.py    — Merchant, MiloCurioDealer, JamboHealsU
    _friends.py      — Mynx, Gorran, Grondite citizens
    _adjutant.py     — TheAdjutant (imports all concrete classes for globals() lookup)

All 46 import dependents continue to resolve unchanged via __init__.py
re-exports. 1 061 tests pass.
