+++ src/npc/_shop.py
